Expanding the asset types/extensions that are targeted by the auto-organiser meant that I had to stop using "OnPostprocessTexture" for the sake of simplicity and testing the capabilities of the AssetPostprocessor class.
To solve my issue of expanding the targeted asset types,  I first applied the OnPostProcessAllAssets function that, per the documentation, is safe to manipulate assets within, including moving them.  This means that I thought I no longer needed the delayCall delegate to register a function to. I was wrong because it seems that this caused memory allocation errors and caused it to work very slowly.I believe due to many operations on it in a quick succession. I re-introduced the delayCall and the issues were resolved.

To solve my issue of not accounting for there being no folder, I first tried to use Sytem.io's Directory.Exists to see if it existed, then created one dynamically as it was needed. This worked for importing assets, but not when assets were created in the engine as it seems the timing was slightly different when creating within the engine instead of importing. Instead of attempting to work around the timing for each, I decided design wise it wasn't necessary or even good to create the folders on the fly as they were needed. Instead I went with initialising a default folder structure on load that is created prior to any importation of assets. This is good also because I eventually want these structures to be customisable by the user eventually.

Next I want to implement the ability to change the folder structure that is applied on the user end.