Previously I had played with the underlying logic of how I map and automatically organise files to folders and for the past week I have been learning how to implement the GUI for a  custom editor window to enable the user to be able to customise the folder structure and naming conventions themselves, save it as a preset and use it. Alternatively the user can use just the default preset structured using Unity's best practice guidelines for project organisation.
This endeavour proved to be more difficult than I had first expected. I had to learn (and I am still learning) about the UI toolkit and the different API's used within it. I found the documentatiion to be vauge and/or confusing in places and it was hard to find definitive answers for how to structure and populate VisualElements such as a TreeView.

In the end I pieced together the various doc pages I needed to understand how to populate the treeview with the appopriate data and the editor window is appearing asexpected.

I removed the CreateDefaultFolders method in FolderStructureManager as the next steps are to create folders according the folder structure defined in within the FolderStructureManager itself , not in AssetFolderSettings
(DefaultFolderStructure initially, and currently thats the only option until I add functionality for creating/saving custom presets to use)

I also deleted the entire AssetFolderSettings because I now will be using the FolderStructureManger and the structures defined within so that im not reusing code and I have  only one place where the folder structure is being defined.