Implemented new GUI elements for the user to be able to customise the mappings between files and folders for automatic asset organisation.
For these GUI elements I have implemented the UI behaviour of the mappings section and the logic and functionality for the delete mapping button.
the user now how the ability to view the associated extensions of selected folders in the treeview and remove them if they wish.

Next step is to implement the Add mapping button logic, then its onto the final stage of applying these presets to the actual project folder structure.